Clean up tidying/setMarks
There is no reason to keep marks beyond the arity of a function, as we don't want to eta expand workers that far.
In fact for join points we are not allowed to eta expand! So it's crucial we don't keep more marks than their (value) arity.

Also during tidy we now get the rhss arguments from the original rhs. Because the tidied one is knot-tied and seems to depend on the binder.
But we have to look at the rhs to fully set the binders marks! So we just set the mark by looking at the original rhs instead.